% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{appServer}
\alias{appServer}
\title{appServer}
\usage{
appServer(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
A shiny app server object
}
\description{
Server logic for the smerte app
}
