% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{bygg_query}
\alias{bygg_query}
\title{bygg_query}
\usage{
bygg_query(registryName, tableName, userInput, nasjonalt_filter)
}
\arguments{
\item{registryName}{Databasenavn for å bygge SQL-query. Hentes automatisk for
user i appServer.}

\item{tableName}{Navn på koblet uttrekk som ønskes. Per nå er det støtte for
'allevarnum', 'smertediagnosernum' og 'smertediagnoser'.}

\item{userInput}{parametre hentet fra bruker-input i shiny.}

\item{nasjonalt_filter}{for å ta bort pasienter som ikke skal overføres til rapporter
på nasjonal instans.}
}
\value{
Returnerer spørring for aktuelt uttrekk.
}
\description{
Hjelpefunksjon for å bygge spørringer for å generere koblede
uttrekk for smerteregisteret. Kjøres som del av datauttrekk til Rapporteket.
Ikke ment å kalles direkte.
}
\keyword{internal}
