% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeRegistryDb.R
\name{describeRegistryDb}
\alias{describeRegistryDb}
\title{Provide registry database metadata}
\usage{
describeRegistryDb(registryName, tabs = c())
}
\arguments{
\item{registryName}{String providing the registry name key}

\item{tabs}{Character vector for optional definition of tables to describe.
Defaults to an empty vector in which case all tables are used}
}
\value{
A list with table names and corresponding fields with attributes
}
\description{
List all tables and fields with attributes such as type and degault values
}
