% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fikseSykehusnavn.R
\name{fikse_sykehusnavn}
\alias{fikse_sykehusnavn}
\title{Legg til sykehusnavn}
\usage{
fikse_sykehusnavn(d, reshIdVar = "UnitId")
}
\arguments{
\item{d}{tibble eller data.frame, som inneholder variabelen `reshIdVar`.}

\item{reshIdVar}{Tekststreng for å identifisere hvilken variabel som inneholder
ReshId for enhetene. Kan varierer mellom ulike tabeller.}
}
\value{
Returnerer opprinnelig datasett i tillegg til kolonnene `orgname`,
`SykehusNavn` og `SykehusKortnavn` hvis disse ikke finnes fra før.
}
\description{
Legger til eller erstatter variablene `orgname`, `SykehusNavn` og
`SykehusKortnavn` hvis disse ikke finnes i datasettet `d`.
}
\examples{
x <- data.frame(UnitId = as.character(c(100089, 4201115, NA, 705758, 4204083)))
x |> fikse_sykehusnavn(reshIdVar = "UnitId")
}
