% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getListTextFunction}
\alias{getListTextFunction}
\title{getListTextFunction}
\usage{
getListTextFunction(registryName)
}
\arguments{
\item{registryName}{Databasekobling.}
}
\description{
Hjelpefunksjon for å trekke ut kategoriske labels fra
'text'-tabell for smerteregisteret. Brukes per nå i uttrekk 'smertediagnoser'.
Åpner kobling mot database via rapbase-funksjoner. Kjøres som del av
datauttrekk til Rapporteket. Ikke ment å kalles direkte.
}
\keyword{internal}
