% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataLokalTilsynsrapportMaaned}
\alias{getRegDataRapportDekningsgrad}
\alias{getRegDataRapportDekningsgradReservasjon}
\alias{getRegDataIndikator}
\alias{getRegDataOpiodReduksjon}
\alias{getSmerteDiagKatValueLab}
\alias{getRegDataSmertekategori}
\alias{getRegDataTimetodeath}
\alias{getRegDataSpinalkateter}
\alias{getRegDataLokalEpidural}
\alias{getRegDataRapportOppfolg}
\alias{getLocalYears}
\alias{getAllYears}
\alias{getHospitalName}
\alias{getNameReshId}
\alias{getDataDump}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataRapportDekningsgrad(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataRapportDekningsgradReservasjon(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataIndikator(registryName, reshId, userRole, startDate, endDate, ...)

getRegDataOpiodReduksjon(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getSmerteDiagKatValueLab(registryName, reshId, smerteKat)

getRegDataSmertekategori(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataTimetodeath(registryName, reshId, userRole, startDate, endDate, ...)

getRegDataSpinalkateter(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataLokalEpidural(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataRapportOppfolg(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getLocalYears(registryName, reshId, userRole)

getAllYears(registryName, reshId, userRole)

getHospitalName(registryName, reshId, userRole)

getNameReshId(registryName, reshId, asNamedList = FALSE)

getDataDump(registryName, reshId, userRole, tableName, fromDate, toDate, ...)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing organization Id}

\item{userRole}{String providing user role}

\item{startDate}{String defing start of date range as YYYY-MM-DD}

\item{endDate}{String defing end of date range as YYYY-MM-DD}

\item{...}{Optional arguments to be passed to the function}

\item{smerteKat}{Integer defining the SmerteKat code to use in query}

\item{asNamedList}{Logical whether to return a list of named values or not.}

\item{tableName}{String providing a table name}

\item{fromDate}{String providing start date}

\item{toDate}{String provideing end date}
}
\description{
Provides a dataframe containing data from a registry
}
