% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNationalReg.R
\name{isNationalReg}
\alias{isNationalReg}
\title{Is organization a national registry?}
\usage{
isNationalReg(reshId)
}
\arguments{
\item{reshId}{Integer organization ID}
}
\value{
Logical (TRUE or FALSE)
}
\description{
By matching organization ID and configuration, TRUE or FALSE is returned
if that organization represents the national registrey or not
}
