% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTenant.R
\name{isTenant}
\alias{isTenant}
\alias{isNationalReg}
\alias{isOUSReg}
\title{Do the current id belong to a given organization?}
\usage{
isNationalReg(reshId)

isOUSReg(reshId)
}
\arguments{
\item{reshId}{Integer organization ID}
}
\value{
Logical (TRUE or FALSE)
}
\description{
By matching organization ID and configuration, TRUE or FALSE is returned
}
