% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubscriptionLocalTilsyn.R
\name{subscriptionLocalTilsyn}
\alias{subscriptionLocalTilsyn}
\title{Provide reports according to subscription}
\usage{
subscriptionLocalTilsyn(baseName, reshId, registryName, year, author,
  hospitalName, type)
}
\arguments{
\item{baseName}{Single element list with value of Rmd template for the report.
Value must be provided without the file extention (\emph{i.e.} ".Rmd")}

\item{reshId}{Single element list with the organization id of the subscriber
as its value}

\item{registryName}{Single element list which value is the regisry name}

\item{year}{Single element list which value is the year used by the report}

\item{author}{Single element list holding the name of the subscriber as its
value}

\item{hospitalName}{Single element list with a (human readable) org name as
its value}

\item{type}{Single element list which value defining report file format,
currently one of "pdf" or "html"}
}
\value{
Full path of file produced
}
\description{
When called, typically through a chain of events initiated by a chron
process, this is the function that actually produces the report to be
shiped off to the recipient. The actual call to this function is made
through do.call and has the effect of providing the parameters as class
\emph{list}. Hence, values must be extracted by list operations.
}
