% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleVariabelOversikt.R
\name{variabelOversiktRapport}
\alias{variabelOversiktRapport}
\alias{variabeloversiktInput}
\title{Module for VariabeloversiktRapport}
\usage{
variabeloversiktInput(
  id,
  startDate = lubridate::today() - lubridate::years(1),
  endDate = lubridate::today() - lubridate::weeks(1),
  min = "1980-01-01",
  max = "2100-01-01"
)
}
\arguments{
\item{id}{Character string with module id (namespace).}

\item{startDate}{Character string of the form YYYY-MM-DD or Date object
providing start date. Default is one year ago.}

\item{endDate}{Character string of the form YYYY-MM-DD or Date object
providing end date. Default is today minus one week.}

\item{min}{Minimum value for startDate and endDate, Default is '1980-01-01'.}

\item{max}{Maximum value for startDate and endDate, Default is '2100-01-01'.}
}
\value{
Shiny object
}
\description{
Module for VariabeloversiktRapport
}
