% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{bygg_query}
\alias{bygg_query}
\title{bygg_query}
\usage{
bygg_query(tableName, userInput)
}
\arguments{
\item{tableName}{Navn på koblet uttrekk som ønskes. Per nå er det støtte for
'allevarnum', 'smertediagnosernum' og 'smertediagnoser'.}

\item{userInput}{parametre hentet fra bruker-input i shiny.}
}
\value{
Returnerer spørring for aktuelt uttrekk.
}
\description{
Hjelpefunksjon for å bygge spørringer for å generere koblede
uttrekk for smerteregisteret. Kjøres som del av datauttrekk til Rapporteket.
Ikke ment å kalles direkte.
}
\keyword{internal}
