% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDefaultReport.R
\name{defaultReport}
\alias{defaultReport}
\alias{defaultReportInput}
\alias{defaultReportUI}
\alias{defaultReportServer}
\alias{defaultReportServer2}
\alias{defaultReportApp}
\title{Shiny modules for Smertereg at Rapportket}
\usage{
defaultReportInput(
  id,
  startDate = today() - years(1),
  endDate = today() - weeks(1),
  min = "1980-01-01",
  max = "2100-01-01"
)

defaultReportUI(id)

defaultReportServer(id, reportFileName, reportParams)

defaultReportServer2(id, reportFileName, reportParams)

defaultReportApp()
}
\arguments{
\item{id}{Character string with module id (namespace).}

\item{startDate}{Character string of the form YYYY-MM-DD or Date object
providing start date. Default is one year ago.}

\item{endDate}{Character string of the form YYYY-MM-DD or Date object
providing end date. Default is today minus one week.}

\item{reportFileName}{Character string providing basename of the file
representing the report template. Such templates must be placed directly
under the inst directory of the package.}

\item{reportParams}{A named list of parameters that will be used by the
report template.}
}
\value{
Shiny object
}
\description{
Shiny modules for Smertereg at Rapportket
}
