% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDefaultReport.R
\name{defaultReport}
\alias{defaultReport}
\alias{defaultReportInput}
\alias{defaultReportUI}
\alias{defaultReportApp}
\alias{defaultReportServer}
\title{Shiny modules for Smertereg at Rapportket}
\usage{
defaultReportInput(
  id,
  startDate = lubridate::today() - lubridate::years(1),
  endDate = lubridate::today() - lubridate::weeks(1)
)

defaultReportUI(id)

defaultReportServer(id, reportFileName, reportParams)

defaultReportApp()
}
\arguments{
\item{id}{Character string with module id (namespace).}

\item{startDate}{Character string of the form YYYY-MM-DD or Date object
providing start date. Default is one year ago.}

\item{endDate}{Character string of the form YYYY-MM-DD or Date object
providing end date. Default is today minus one week.}
}
\value{
Shiny object
}
\description{
Shiny modules for Smertereg at Rapportket
}
