% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fikse_sykehusnavn.R
\name{fikse_sykehusnavn}
\alias{fikse_sykehusnavn}
\title{Add identifying names to the UnitId}
\usage{
fikse_sykehusnavn(df, reshIdVar = UnitId)
}
\arguments{
\item{df}{data.frame, must contain variable `reshIdVar`}

\item{reshIdVar}{Variable identifying the ReshId for the units. Can have
different names depending on the table selected.}
}
\value{
data.frame with variable `orgname`. Old values are overwritten
if `orgname` already existed if `df`
}
\description{
Add variables `orgname`, `SykehusNavn` and
`SykehusKortnavn` if missing, update values if existing.
}
\examples{
x <- data.frame(UnitId = as.character(c(100089, 4201115, NA, 705758, 4204083)))
x |> fikse_sykehusnavn(reshIdVar = "UnitId")

}
