% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fikse_sykehusnavn.R
\name{fikse_sykehusnavn}
\alias{fikse_sykehusnavn}
\title{Add variable `orgname` if missing, update values if existing}
\usage{
fikse_sykehusnavn(df)
}
\arguments{
\item{df}{data.frame, must contain variable `UnitId`}
}
\value{
data.frame with variable `orgname`. Old values are overwritten
if `orgname` already existed if `df`
}
\description{
Add variable `orgname` if missing, update values if existing
}
\examples{
x <- data.frame(UnitId = c(108141, 109880, NA, 123, 105502))
x \%>\% fikse_sykehusnavn()

y <- data.frame(UnitId = c(108141, 109880, NA, 123, 105502),
                orgname = c("AHUS", "Ullevål", "NA",
                 "test", "Stavanger"))
}
