% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataLokalTilsynsrapportMaaned}
\alias{getRegDataRapportDekningsgrad}
\alias{getLocalYears}
\alias{getHospitalName}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(registryName, reshId, userRole, year, ...)

getRegDataRapportDekningsgrad(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getLocalYears(registryName, reshId, userRole)

getHospitalName(registryName, reshId, userRole)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing organization Id}

\item{userRole}{String providing user role}

\item{year}{Integer four digit year to be reported from}

\item{...}{Optional arguments to be passed to the function}

\item{startDate}{String defing start of date range as YYYY-MM-DD}

\item{endDate}{String defing end of date range as YYYY-MM-DD}
}
\description{
Provides a dataframe containing data from a registry
}
