% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R, R/GetRegDataOld.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataTilsynsrapportMaaned}
\alias{getLocalYears}
\alias{getRegDataLokalTilsynsrapportMaaned}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(registryName, reshId, startDate,
  endDate)

getLocalYears(registryName, reshId)

getRegData(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing reshId}

\item{startDate}{Start date ...}

\item{endDate}{End date...}

\item{registryName}{String providing the current registryName}
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry

Provides a dataframe containing data from a registry
}
