% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataLokalTilsynsrapportMaaned}
\alias{getRegDataRapportDekningsgrad}
\alias{getLocalYears}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(registryName, year, ...)

getRegDataRapportDekningsgrad(registryName, reshId, startDate, endDate,
  ...)

getLocalYears(registryName, reshId)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{year}{Integer four digit year to be reported from}

\item{...}{Optional arguments to be passed to the function}

\item{reshId}{String providing organization Id}

\item{startDate}{String defing start of date range as YYYY-MM-DD}

\item{endDate}{String defing end of date range as YYYY-MM-DD}
}
\description{
Provides a dataframe containing data from a registry
}
