% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataLokalTilsynsrapportMaaned}
\alias{getRegDataRapportDekningsgrad}
\alias{getRegDataSmertekategori}
\alias{getRegDataSpinalkateter}
\alias{getLocalYears}
\alias{getAllYears}
\alias{getHospitalName}
\alias{getDataDump}
\alias{getRegDataIndikator}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(registryName, reshId, userRole, year, ...)

getRegDataRapportDekningsgrad(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataIndikator(registryName, reshId, userRole, year, ...)

getRegDataSmertekategori(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getRegDataSpinalkateter(
  registryName,
  reshId,
  userRole,
  startDate,
  endDate,
  ...
)

getLocalYears(registryName, reshId, userRole)

getAllYears(registryName, reshId, userRole)

getHospitalName(registryName, reshId, userRole)

getDataDump(registryName, tableName, fromDate, toDate, ...)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing organization Id}

\item{userRole}{String providing user role}

\item{year}{Integer four digit year to be reported from}

\item{...}{Optional arguments to be passed to the function}

\item{startDate}{String defing start of date range as YYYY-MM-DD}

\item{endDate}{String defing end of date range as YYYY-MM-DD}

\item{tableName}{String providing a table name}

\item{fromDate}{String providing start date}

\item{toDate}{String provideing end date}
}
\description{
Provides a dataframe containing data from a registry
}
