% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataLokalTilsynsrapportMaaned}
\alias{getLocalYears}
\title{Provide dataframe of registry data}
\usage{
getRegDataLokalTilsynsrapportMaaned(registryName, reshId, userRole, year, ...)

getLocalYears(registryName, reshId)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing organization Id}

\item{year}{Integer four digit year to be reported from}

\item{...}{Optional arguments to be passed to the function}
}
\description{
Provides a dataframe containing data from a registry
}
