% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_behandlertilsyn}
\alias{ki_behandlertilsyn}
\title{Kvalitetsindikator behandlertilsyn}
\usage{
ki_behandlertilsyn(d)
}
\arguments{
\item{d}{datasett som inneholder de radene vi skal beregne indikator for.}
}
\value{
Returnerer Inndatasett med følgende ekstra variabler.

\item{ki_krit_teller}{TRUE/FALSE avhengig av om pasienten skal være med i teller.}
\item{ki_krit_nevner}{TRUE/FALSE avhengig av om pasienten skal være med i nevner.}
}
\description{
Indikatoren regner ut andel pasienter som mottar tilsyn fra to eller flere
behandlergrupper.
Dette finnes ved å telle opp antall registrerte tilsyn for
Lege (`AntTilsLege`), Sykepleier (`AntTilsSykPleier`),
Fysioterapeut (`AntTilsFysioT`), Psykolog (`AntTilsPsyk`),
Sosionom (`AntTilsSosio`), og Konferert med lege (`AntTilsKonfLege`).

Funksjonen teller opp antall registrerte tilsyn for overnevnte variabler
og returnerer `ki_krit_nevner = TRUE/FALSE` avhengig av om pasienten skal
inkluderes i indikator.
`ki_krit_teller = TRUE/FALSE` avhengig av om det er registrert to eller flere
tilsyn og `ki_krit_nevner = TRUE`.

Utdata kan videre mates inn i `rapwhale::aggreger_ki_prop()` for å beregne
andel for aktuelle undergrupper.
}
\examples{
d_behtils = ki_behandlertilsyn(d)
}
