% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodebokKategoriske.R
\name{lag_smerte_kb_kategoriske}
\alias{lag_smerte_kb_kategoriske}
\title{Lag kodebok for kategoriske variabler i smerteregisteret.}
\usage{
lag_smerte_kb_kategoriske(d_text)
}
\arguments{
\item{d_text}{datasett som inneholder verditekster for listeverdiene.}
}
\value{
Returnerer en tibble med kategoriske variabler for alle tabeller oppgitt i
'tabeller'-objektet i funksjonen. Denne kan redigeres hvis nødvendig.
}
\description{
Trekker ut verditekst for kategoriske variabler fra 'text'-tabellen i
databasen. Utdata lagres på format som er forenelig med 'rapwhale::kb_fyll()'
og har variablene `tabell`, `variabel_id`, `verdi` og `verditekst`.

Vær obs på at du kobler mot riktig tabell da jeg ikke har sjekket om alle
kategoriske variabler har samme verdi-verditekst kombinasjon på tvers av skjema.
}
