% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeRegistryName.R
\name{makeRegistryName}
\alias{makeRegistryName}
\title{Provide registry name corresponding to db config entry}
\usage{
makeRegistryName(baseName, reshID)
}
\arguments{
\item{baseName}{String giving the prefix base of the name}

\item{reshID}{String providing the current reshID. At Rapporteket, reshID
should already be present in the current R session}
}
\value{
String containing a valid registry name
}
\description{
Provides the registry name key to be used for reading correspponding
entries from the (yaml) configuration file
}
