% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerte-package.R
\docType{package}
\name{smerte-package}
\alias{smerte}
\alias{smerte-package}
\title{smerte: R-pakke for smerteregisteret}
\description{
Pakke for det nasjonale registeret "Smerteregisteret". Utarbeiding av rapporter for lokale og nasjonale data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rapporteket/smerte}
  \item Report bugs at \url{https://github.com/rapporteket/smerte/issues}
}

}
\author{
\strong{Maintainer}: Janne Mannseth \email{janne.mannseth@helse-bergen.no}

Authors:
\itemize{
  \item Janne Mannseth \email{janne.mannseth@helse-bergen.no}
  \item Are Edvardsen \email{biorakel@gmail.com} (\href{https://orcid.org/0000-0002-5210-3656}{ORCID})
  \item Endre Kvåle Evjen \email{endre.kvale.evjen@helse-bergen.no}
  \item Kevin Otto Thon \email{kevin.thon@helse-nord.no}
  \item Arnfinn Hykkerud Steindal \email{arnfinn.steindal@gmail.com} (\href{https://orcid.org/0000-0002-4088-7860}{ORCID})
}

}
\keyword{internal}
