% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smerte-package.R
\docType{package}
\name{smerte-package}
\alias{smerte}
\alias{smerte-package}
\title{smerte: R-pakke for smerteregisteret}
\description{
Pakke for det nasjonale registeret "Smerteregisteret". Utarbeiding av rapporter for lokale og nasjonale data.
}
\author{
\strong{Maintainer}: Janne Mannseth \email{janne.mannseth@helse-bergen.no}

Authors:
\itemize{
  \item Are Edvardsen \email{biorakel@gmail.com} (\href{https://orcid.org/0000-0002-5210-3656}{ORCID})
}

}
\keyword{internal}
